Removed the common package
