--- packages/common/.flowconfig
@@ -1,7 +1,0 @@
-[ignore]
-
-[include]
-
-[libs]
-
-[options]
