--- packages/common/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "packages",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "private": true,
-  "devDependencies": {
-    "babel-plugin-external-helpers": "^6.22.0"
-  }
-}
