+++ packages/container-query/.flowconfig
@@ -2,7 +2,6 @@
 ./dist
 
 [include]
-../common/src
 
 [libs]
