+++ packages/container-query/src/getConditionFunction.js
@@ -1,5 +1,5 @@
 // @flow
-import type { ContainerSize } from "../../common/src/types";
+import type { ContainerSize } from "../flow/types";
 
 /**
  * Combines condition functions into a single "and" one.
