+++ packages/demos/nested-containers/src/css/components/user/user.json
@@ -1,1 +1,1 @@
-{"selector":".user","queries":[{"elements":[{"selector":".user","values":{"border":"0.5rmax solid","borderRadius":"3rmin"}},{"selector":".user__info","values":{"padding":"1rmax"}},{"selector":".user__image","values":{"maxWidth":"100rw","height":"30rw","width":"30rw","borderWidth":"1.25rmin"}}]},{"conditions":[[["width",">",300]]],"elements":[{"selector":".user__info","styles":{"flexDirection":"row"}}]},{"conditions":[[["width",">",100]]],"elements":[{"selector":".user__name","styles":{"display":"block"}}]},{"conditions":[[["width",">",300]]],"elements":[{"selector":".user__name","styles":{"textAlign":"left","width":"auto","margin":"0"},"values":{"paddingLeft":"5rw"}}]}]}
+{"selector":".user","queries":[{"elements":[{"selector":".user","values":{"border":"0.5rmax solid","borderRadius":"3rmin"}},{"selector":".user__info","values":{"padding":"1rmax"}},{"selector":".user__image","values":{"maxWidth":"100rw","height":"30rw","width":"30rw","borderWidth":"1.25rmin"}}]},{"conditions":[[["width",">",300]]],"elements":[{"selector":".user__info","styles":{"webkitBoxOrient":"horizontal","webkitBoxDirection":"normal","msFlexDirection":"row","flexDirection":"row"}}]},{"conditions":[[["width",">",100]]],"elements":[{"selector":".user__name","styles":{"display":"block"}}]},{"conditions":[[["width",">",300]]],"elements":[{"selector":".user__name","styles":{"textAlign":"left","width":"auto","margin":"0"},"values":{"paddingLeft":"5rw"}}]}]}
